==> src/3_eda_stats_fotmob.ipynb <==
import eredivisie_nlp as enlp
import os
import pandas as pd

# load data
df = pd.read_pickle(f"{enlp.determine_root()}/data/raw/grades/fotmob.pkl")

## Player stats

# summary numbers on the player stats
df_stats = df.drop(['hashtag', 'name', ' FotMob rating'], axis=1)

perc_non_na = (1 - (df_stats.isna().sum().sum() / (df_stats.shape[0] * df_stats.shape[1]))) * 100
print(f"{perc_non_na}% of the dataframe is data.")
print(f"That means we have on average, {len(df_stats.columns) * 0.44698} fields per player.")
del(df_stats)
