==> src/4_3_train_test_split.ipynb <==
import eredivisie_nlp as enlp
import pandas as pd
from sklearn.model_selection import train_test_split

df = pd.read_excel(enlp.determine_root() + "/data/sentiment_annotations_manual_no_leakage.xlsx")
df = df[df.annotation != 100]
train, test = train_test_split(df, test_size=100, random_state=237)

train = train.reset_index(drop=True)
test = test.reset_index(drop=True)
train.to_csv(enlp.determine_root() + "/data/train_set.csv")
test.to_csv(enlp.determine_root() + "/data/test_set.csv")
