==> src/6_1_results_rq2.ipynb <==
import eredivisie_nlp as enlp
import numpy as np
import pandas as pd
from pattern.nl import sentiment
from random import randrange
from sklearn.metrics import mean_absolute_error, cohen_kappa_score
from tqdm import tqdm
from transformers import RobertaTokenizer, RobertaForSequenceClassification

# Pattern

test = pd.read_csv(enlp.determine_root() + "/data/test_set.csv", index_col=[0])

Pattern <> manual annotations

test['sent_pattern'] = [sentiment(text)[0] for text in test.text]
test

mean_absolute_error(test.annotation, test.sent_pattern)

Let's compare this with a naive approach

test['naive_6d'] = [randrange(1000000)/1000000 for p in range(test.shape[0])]  # pattern gives 6 decimals
test['naive_1d'] = [randrange(21) / 10 - 1 for x in range(test.shape[0])]  # standard SA is 1 decimal

mean_absolute_error(test.annotation, test.naive_6d)

mean_absolute_error(test.annotation, test.naive_1d)

The number of digits does not have an influence on the performance of a naive classifier
Errors of 0.3 vs 0.6 are less bad than -0.3 vs 0.3. Let's see how often pattern gets the polarity wrong.

# percentage agreement manual <> pattern
agreement = [1 if enlp.is_positive(p) == enlp.is_positive(m) else 0 for p, m in zip(test.sent_pattern, test.annotation)]
sum(agreement) / test.shape[0]

# cohens kappa manual <> pattern
pd.Series([enlp.is_positive(m) for m in test.annotation]).value_counts()
cohen_kappa_score([enlp.is_positive(a) for a in test.annotation], [enlp.is_positive(p) for p in test.sent_pattern])

Now when transforming pattern.nl annotations to the same format as the manual ones.

test['pattern_round'] = [enlp.round_score(p) for p in test.sent_pattern]

mean_absolute_error(test.annotation, test.pattern_round)

test['annotation_pol'] = [enlp.is_positive(a) for a in test.annotation]
test['pattern_pol'] = [enlp.is_positive(p) for p in test.sent_pattern]
mean_absolute_error(test.annotation_pol, test.pattern_pol)

# RobBERT

# load models and tokenizer
models = {
    'oneshot': RobertaForSequenceClassification.from_pretrained("pdelobelle/robbert-v2-dutch-base"),
    '150': RobertaForSequenceClassification.from_pretrained("./robbert_150"),
    '300': RobertaForSequenceClassification.from_pretrained("./robbert_300"),
    '600': RobertaForSequenceClassification.from_pretrained("./robbert_600")
}
tokenizer = RobertaTokenizer.from_pretrained("pdelobelle/robbert-v2-dutch-base")

# SA process
model_choice = '600'
model = models[model_choice]

encoded_tweets = [tokenizer(tweet, return_tensors='pt') for tweet in test.text]
outputs = [model(**et)[0].detach().numpy() for et in tqdm(encoded_tweets)]
sentiments = []
for output in outputs:
    robbert_score = np.where(output[0] == max(output[0]))[0].item()
    sentiments.append(enlp.transform_score(robbert_score))
test[f'robbert_{model_choice}'] = sentiments

mean_absolute_error(test.annotation, test[model_choice])

robbert_pol = []
for rp in test['robbert_600']:
    if rp > 0:
        robbert_pol.append(1)
    elif rp < 0:
        robbert_pol.append(-1)
    else:
        robbert_pol.append(0)

test['robbert_pol'] = robbert_pol

pattern_agree = [1 if pp == ap else 0 for pp, ap in zip(test.pattern_pol, test.annotation_pol)]
sum(pattern_agree)

robbert_agree = [1 if rp == ap else 0 for rp, ap in zip(test.robbert_pol, test.annotation_pol)]
sum(robbert_agree)
